Refactor with getDefaultDataSourceId$
